dom/tests/mochitest/ajax/offline/test_bug474696.html

<html xmlns="http://www.w3.org/1999/xhtml" manifest="https://example.com/tests/dom/tests/mochitest/ajax/offline/fallback.cacheManifest">
<head>
<title>Fallback on different origin redirect test</title>

<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script type="text/javascript" src="/tests/dom/tests/mochitest/ajax/offline/offlineTests.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script class="testbody" type="text/javascript">
/* globals fallbackFrame */

function manifestUpdated()
{
  fallbackFrame.location = "https://example.com/tests/dom/tests/mochitest/ajax/offline/namespace1/redirectToDifferentOrigin.sjs";
  // Invokes load of fallback.html
}

function onFallbackLoad(fallbackIdentification)
{
  OfflineTest.is(fallbackIdentification, 1, "Got correct fallback for namespace1 (2)");

  applicationCache.onerror = function() {}; // the update invoked by the iframe will finish after we discard the cache, ignore error
  OfflineTest.teardownAndFinish();
}

SimpleTest.waitForExplicitFinish();

if (OfflineTest.setup()) {
  applicationCache.onerror = OfflineTest.failEvent;
  applicationCache.onupdateready = OfflineTest.failEvent;
  applicationCache.oncached = OfflineTest.priv(manifestUpdated);
}

</script>

</head>

<body>
<iframe name="fallbackFrame"></iframe>
</body>
</html>
